moved filter components in their own component folder, changed some minor things in MUI theme, made the request list header and body more responsive
