+++ src/main.jsx
@@ -26,7 +26,7 @@ const theme = createTheme({
       background: "rgb(62, 62, 62)",
     },
     text: {
-      primary: "rgba(0, 0, 0, 0)",
+      primary: "rgb(0, 0, 0)",
       secondary: "rgb(255, 255, 255)",
     },
   },
@@ -35,10 +35,10 @@ const theme = createTheme({
   },
   breakpoints: {
     values: {
-      sm: 320,
-      md: 440,
-      lg: 767,
-      xl: 1024,
+      xs: 320,
+      sm: 440,
+      md: 767,
+      lg: 1024,
     },
   },
 });
